Product finish, start with Supplier

+++ src/main/java/com/example/ims/models/Supplier.java
@@ -2,23 +2,23 @@ package com.example.ims.models;
 
 public class Supplier {
 
-    private Long supplier_id;
+    private Integer supplier_id;
     private String name;
     private String contactInf;
-    private int phoneNumber;
+    private Integer phoneNumber;
 
-    public Supplier(Long supplier_id, String name, String contactInf, int phoneNumber) {
+    public Supplier(Integer supplier_id, String name, String contactInf, int phoneNumber) {
         this.supplier_id = supplier_id;
         this.name = name;
         this.contactInf = contactInf;
         this.phoneNumber = phoneNumber;
     }
 
-    public Long getSupplier_id() {
+    public Integer getSupplier_id() {
         return supplier_id;
     }
 
-    public void setSupplier_id(Long supplier_id) {
+    public void setSupplier_id(Integer supplier_id) {
         this.supplier_id = supplier_id;
     }
